add batch_assort_mixmat function
Assortativity and mixing matrix elements can now be computed for each
specific groups (i.e. per patient / sample / chromosome, ...), with
number of nodes and proportions of catogories per group.
A parallel version is implemented with Dask, but for few samples and
randomization rounds it's slower than the single cpu version.